pyproject.toml
[tool.poetry]
name = "pybedrockllamaindex-202403"
version = "0.1.0"
description = ""
authors = ["Shuichi Kataoka <shu01k9@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
boto3 = ">=1.28.84"
botocore = ">=1.31.84"
openai = "0.28"
llama-index = "0.8.42"
langchain = "0.0.303"
pypdf = ">=3.17.0"
python-dotenv = "^1.0.1"
streamlit = "^1.32.2"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
